+++ server/src/main/java/replicant/server/ee/ReplicantSystem.java
@@ -1,0 +1,18 @@
+package replicant.server.ee;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import javax.inject.Qualifier;
+
+/**
+ * Qualifier for the replicant system.
+ */
+@Qualifier
+@Retention( RetentionPolicy.RUNTIME )
+@Target( { ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER, ElementType.TYPE } )
+public @interface ReplicantSystem
+{
+  String value() default "";
+}